Fix Py26 compat for 'with gzip.open'.
